backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
python-decouple==3.8
dj-database-url==2.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1
langchain-openai==0.0.8
langgraph==0.0.26
pytz==2023.3
requests[socks]==2.31.0
pysocks==1.7.1
python-dotenv==1.0.0
openai>=1.6.1 